Running with Docker Fix
There were issues with using the environment variables within the default-env.sh file with Docker. I have created a separate file for docker to contain these environment variables and adjusted the scripts to run with Docker locally.

The ReadMe has been updated to reflect these changes.

typeformClient.js was missing an import in order to function correctly - the rest of this file is just stylistic changes.