Fixed Linting Errors
This project uses standard.js as a linting tool. Most of these changes are linting errors that standard.js picked up and fixed to coincide with its standard.

Although passing build and tests locally through Docker, when building through Docker once deployed to the AWS pipeline standard.js was preventing the build to complete due to picking up the syntax for handlebars within the jsreport system as invalid. I've searched around and could not find any way to adjust standard.js to accept this, nor could I find a plugin to help.

I've therefore updated the standard.js settings within the package.json file to ignore the one file that contains this syntax in order for jsreport to build the pdf.