Restore filtering of Hubspot Data
I had mistakenly took out functionality to filter the Hubspot Data retrieved to get the correct user details. Instead I was just returning the first item in the array of Hubspot data returned. Now, the Hubspot Data is filtered to return the entry with a corresponding uuid from the form submission.