Added a separate initial call for Typeform Data.
Originally the buildReportViewModalFor function would be run once to get just the Typeform Data to pass on to the Hubspot Landing page, and then again to render the report with Hubspot Data once the user had submitted their information on the Hubspot landing page.

I've changed this to originally only call for the Typeform Data to send to the Hubspot page, and then call for all data when building the report within jsreport.