Remove unused hardcoded iframe
This iframe is supposed to hold the survey, it
is what's proposed when using the 'share'
option on typeform.

But then, because the 'typeform-survey' div
was not closed, the iframe was inside it. And
when calling 'typeformEmbed.makeWidget'
on the 'client/index.js' it would override it.
It turns out that this was the intended
behavior anyway, because if it wasn't
overriden, we would have 2 iframes, one
hardcoded in the 'index.ejs' file, and another
one dynamically set by 'typeformEmbed.makeWidget'

So I removed the hardcoded one, and closed the div.