Remove AWS Keys from ElasticBeanstalk environment
The AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY were set in the ElasticBeanstalk environment as
configuration variables, but they are not needed. In fact, on AWS we use a IAM Role, wrapped in an
instance profile, to give the EC2 instance access to other services (S3 & SES).
The keys, however, are needed when running the service locally.

Co-authored-by: Cherry Tsui <cherry.tsui@codurance.com>