Remove un-necessary step in 'create-env'
The deleted step was assigning a permission to our EC2
instance that was not needed and was not meant for
EC2 instances in the first place.
The 'AWSElasticBeanstalkService' permission is a permission that is set on the ElasticBeanstalk service
via the role 'aws-elasticbeanstalk-service-role'. This role
is automatically set by EB on the environment when
creating the environment, so we also removed it from the
'option-settings.json' file to avoid confusion.

Co-authored-by: Cherry Tsui <cherry.tsui@codurance.com>